pyproject.toml
[tool.poetry]
name = "betting_parser"
version = "1.0.0"
description = "Parse data for betting predictions"
authors = ["Alexey Kryuchkov <alexkryu4kov@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.9"

beautifulsoup4 = "4.10.0"
soupsieve = "2.3.1"

pandas = "1.4.1"
numpy = "1.22.2"
pytz = "2021.3"
six = "1.16.0"
python-dateutil = "2.8.2"

selenium = "4.1.0"
pycparser = "2.21"
sniffio = "1.2.0"
trio-websocket = "0.9.2"
trio = "0.20.0"
outcome = "1.1.0"
contextvars = "2.4"
cffi = "1.15.0"
sortedcontainers = "2.4.0"
async-generator = "1.10"
wsproto = "1.0.0"
idna = "3.3"
attrs = "21.4.0"

requests = "2.27.1"
charset-normalizer = "2.0.12"
certifi = "2021.10.8"
chardet = "4.0.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"